+++ project/app/src/androidTest/kotlin/org/owntracks/android/e2e/ContactsAndMapTests.kt
@@ -20,6 +20,8 @@ import org.owntracks.android.testutils.di.setLocation
 import org.owntracks.android.testutils.grantMapActivityPermissions
 import org.owntracks.android.testutils.setNotFirstStartPreferences
 import org.owntracks.android.testutils.use
+import org.owntracks.android.testutils.waitUntilViewDisplayed
+import org.owntracks.android.testutils.waitUntilViewNotDisplayed
 import org.owntracks.android.ui.map.MapActivity
 
 @LargeTest
@@ -57,13 +59,14 @@ class ContactsAndMapTests :
     clickOn("aa")
     waitUntilActivityVisible(MapActivity::class.java)
 
+    waitUntilViewDisplayed(R.id.bottomSheetLayout)
     assertDisplayed(R.id.bottomSheetLayout)
     assertDisplayed(R.id.contactPeek)
     assertContains(R.id.name, "aa")
 
     clickBack()
 
-    assertNotDisplayed(R.id.bottomSheetLayout)
+    waitUntilViewNotDisplayed(R.id.bottomSheetLayout)
     assertNotDisplayed(R.id.contactPeek)
   }
 }