+++ project/app/src/main/java/org/owntracks/android/App.kt
@@ -103,8 +103,7 @@ open class BaseApp :
     Timber.plant(TimberInMemoryLogTree(BuildConfig.DEBUG))
 
     if (BuildConfig.DEBUG) {
-      System.setProperty(org.slf4j.simple.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "debug")
-      org.slf4j.LoggerFactory.getLogger(this::class.java).trace("SLF4J logging at trace level")
+
       Timber.e("StrictMode enabled in DEBUG build")
       StrictMode.setThreadPolicy(
           StrictMode.ThreadPolicy.Builder()